rename plugin metadata files to folow best practices
