+++ packages/forklift-console-plugin/plugin-extensions.ts
