+++ packages/forklift-console-plugin/plugin-metadata..ts
